feat: Add FontAwesome icons and SkillsComponent
The code changes include adding the "@fortawesome/fontawesome-free" dependency and importing the corresponding CSS file in "main.js" and "App.vue". Additionally, a new route and component, "SkillsComponent", have been added to the router configuration.

This commit message follows the established convention of using a prefix to indicate the type of change ("feat" for a new feature) and provides a clear and concise summary of the changes made.