refactor: Update header information in HeaderComponent.vue
The code changes in this commit update the header information in the HeaderComponent.vue file. The name has been changed to "Adrian Hinojosa", the title has been changed to "Developer", and the email address has been updated to "adrian.armando20@gmail.com". This change reflects the correct information for the header component.

This commit message follows the established convention of using a prefix to indicate the type of change ("refactor" for code refactoring) and provides a clear and concise summary of the changes made.